Added clock configuration options
Default clock now configured to 100MHz (so that we have 10ns increments).
Added serial command to reconfigure clock on the fly
Serial command also supports switching to external clock input on GPIO 20 or 22 (and a 48MHz clock is output on GPIO 21 for debugging this - aka, it can be easily jumpered to GPIO 20)

TODO: RP2040 datasheet says using resus trigger to switch clocks is not meant for production. But I can't get it to switch clk_sys clocks correctly outside of resus!
Also, resus is not firing if running from an external clock and that clock stops. In that scenario, the entire board locks up (except the PIO core keeps executing, but really slow and may get stuck repeating until a clock is restored - but the serial input is still locked up...needs more investigating!)