Fixed a couple of minor bugs in how waits are reported.
The first is related to indefinite waits. The length of time it takes between timeout and beginning next instruction is 1 clock cycle longer than expected so now we report it as 4294967295 to distinguish it.

The second relates to the changes in #12 - I did not update the wait length offset to account for the reduction in minimum half period.