Experimental firmware synchronising PIO state machines
Following the information in aspberrypi/pico-feedback#65 I have changed the code that starts the PIO state machines to also synchronise the internal clocks.

I'm not sure if this will actually do anything because we're not running them with different clock dividers (or any divider as far as I'm aware!) but I guess we will see!